fixes, b64 QVA/QPA

+++ shaders/main.glsl
@@ -13,8 +13,10 @@
 //#define USE_PASS_AO 1
 
 #ifdef GL_ES
-precision mediump float;
-precision mediump int;
+//precision mediump float;
+//precision mediump int;
+precision highp float;
+precision highp int;
 #endif
 
 #define PI     3.1415926535897932
